OT167-23: reanmed and moved mapper folder and rename method
